Separate component for Profile Picture.
